requirements.txt
pandas
numpy==1.24.4
scikit-learn
scikit-surprise
streamlit